+++ src/serlio/prtModifier/MayaCallbacks.h
@@ -82,7 +82,7 @@ public:
 		double const* const* uvs, size_t const* uvsSizes,
 		uint32_t const* const* uvCounts, size_t const* uvCountsSizes,
 		uint32_t const* const* uvIndices, size_t const* uvIndicesSizes,
-		uint32_t uvSets,
+		size_t uvSets,
 		const uint32_t* faceRanges, size_t faceRangesSize,
 		const prt::AttributeMap** materials,
 		const prt::AttributeMap** reports,
